[Refactor] 커스텀 컴포넌트 파일 분리
