+++ TILApp/Views/Shared/Components/CustomFollowButton.swift
@@ -1,0 +1,35 @@
+//
+//  CustomFollowButton.swift
+//  TILApp
+//
+//  Created by 이재희 on 10/27/23.
+//
+
+import FlexLayout
+import PinLayout
+import Then
+import UIKit
+
+class CustomFollowButton: UIButton {
+    override init(frame: CGRect) {
+        super.init(frame: frame)
+        setupButton()
+    }
+
+    required init?(coder: NSCoder) {
+        super.init(coder: coder)
+        setupButton()
+    }
+
+    private func setupButton() {
+        backgroundColor = UIColor(named: "AccentColor")
+        layer.cornerRadius = 12
+        setTitleColor(UIColor(white: 1.0, alpha: 1.0), for: .normal)
+        titleLabel?.font = UIFont.boldSystemFont(ofSize: 13)
+    }
+
+    override func layoutSubviews() {
+        super.layoutSubviews()
+        pin.width(90).height(30)
+    }
+}
