+++ TILApp/Views/Shared/Components/CustomTextField.swift
@@ -1,0 +1,29 @@
+//
+//  CustomTextField.swift
+//  TILApp
+//
+//  Created by 이재희 on 10/27/23.
+//
+
+import FlexLayout
+import PinLayout
+import Then
+import UIKit
+
+class CustomTextField: UITextField {
+    override init(frame: CGRect) {
+        super.init(frame: frame)
+        borderStyle = .roundedRect
+        font = UIFont.systemFont(ofSize: 15, weight: .regular)
+    }
+
+    @available(*, unavailable)
+    required init?(coder _: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+
+    override func layoutSubviews() {
+        super.layoutSubviews()
+        pin.horizontally(20).height(40)
+    }
+}
