+++ TILApp/Views/Shared/Components/CustomTextFieldView.swift
@@ -1,0 +1,68 @@
+//
+//  CustomTextFieldView.swift
+//  TILApp
+//
+//  Created by 이재희 on 10/27/23.
+//
+
+import FlexLayout
+import PinLayout
+import Then
+import UIKit
+
+class CustomTextFieldView: UIView {
+    private let titleLabel = CustomTitleLabel()
+    private let textField = CustomTextField()
+
+    private let height: CGFloat = 64
+    var componentSize: CGSize {
+        return CGSize(width: frame.width, height: height)
+    }
+
+    var titleText: String {
+        get { titleLabel.text ?? "" }
+        set { titleLabel.text = newValue }
+    }
+
+    var placeholder: String {
+        get { textField.placeholder ?? "" }
+        set { textField.placeholder = newValue }
+    }
+
+    var mainText: String {
+        get { textField.text ?? "" }
+        set { textField.text = newValue }
+    }
+
+    var delegate: UITextFieldDelegate? {
+        get { textField.delegate }
+        set { textField.delegate = newValue }
+    }
+
+    override init(frame: CGRect) {
+        super.init(frame: frame)
+
+        setupView()
+    }
+
+    @available(*, unavailable)
+    required init?(coder _: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+
+    private func setupView() {
+        // backgroundColor = .systemBackground
+
+        flex.define {
+            $0.addItem(titleLabel).margin(0, 20).height(24)
+            $0.addItem(textField).margin(0, 20).height(40)
+        }
+    }
+
+    override func layoutSubviews() {
+        super.layoutSubviews()
+
+        pin.width(100%).height(height)
+        flex.layout()
+    }
+}
