+++ TILApp/Views/Shared/Components/CustomTitleLabel.swift
@@ -1,0 +1,29 @@
+//
+//  CustomTitleLabel.swift
+//  TILApp
+//
+//  Created by 이재희 on 10/27/23.
+//
+
+import FlexLayout
+import PinLayout
+import Then
+import UIKit
+
+class CustomTitleLabel: UILabel {
+    override init(frame: CGRect) {
+        super.init(frame: frame)
+        font = UIFont.systemFont(ofSize: 14, weight: .semibold)
+        textColor = UIColor.systemGray
+    }
+
+    @available(*, unavailable)
+    required init?(coder _: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+
+    override func layoutSubviews() {
+        super.layoutSubviews()
+        pin.height(24)
+    }
+}
